Removing compiler warnings

+++ apps/OpenSpace/ext/launcher/src/profile/assetsdialog.cpp
@@ -207,7 +207,7 @@ QString AssetsDialog::createTextSummary() {
         return "";
     }
     QString summary;
-    for (int i = 0; i < summaryItems.size(); ++i) {
+    for (size_t i = 0; i < summaryItems.size(); ++i) {
         bool existsInFilesystem = summaryItems.at(i)->doesExistInFilesystem();
 
         std::string s = existsInFilesystem ?
